requirements.txt
alembic==1.15.2
Flask==3.1.0
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.22.0
langchain_core==0.3.56
langchain_openai==0.3.14
openai==1.76.0
pydantic==2.11.3
pytest==8.3.5
python-dotenv==1.1.0
WTForms==3.2.1
